Add style to index, fried-rice pages
I would like to make the fried rice a vignette but perhaps I'm
encountering a moment when making the image correct in a
paint program is the actual correct thing to do. I recall that
in maybe my network fundamentals course? Somewhere.

It's hideous, but it's mine.

I'm unsure why but the parent folder odin-recipes is not considered
a git tracked repository or something. I have to manually add
the files I want rather than adding the whole folder for a commit.